requirements.txt
streamlit==1.7.0
torch==1.11.0
torchvision==0.12.0
Pillow==7.1.2
matplotlib==3.2.2
